test(openmemory-js): port ad-hoc tsx scripts to vitest specs
- multilingual_dedup: move out of folder-named-like-a-file,
  rewrite as vitest describe/it preserving original assertions
- omnibus: convert phases 1-3 to it() blocks; keep mocked Date.now
  for the evolutionary-stability phase, preserve sleeps and cleanup
- verify: quarantine via describe.skip with an explanatory TODO.
  Original used q.conn.run (not in current db.ts API) and asserted
  hard-coded sector classifications that depend on classifier
  heuristics — re-baselining is out of scope for this infra pass.
  Cleanup-API and embedding-key issues are pre-fixed in the .skip
  body so a future implementer only needs to revisit the assertions.
